Add JavaScript functionality to show data by making requests to the APIs
Now, the 'programming languages select' dropdown is populated with data from the API. When a programming language is selected from the dropdown, a request is made to the GitHub API, and the data is stored in variables. I also added a function to get random indices so that a random repository can be selected.